Resolve "create command line tools to train."

Closes #14

See merge request romain.le-miere/ocr-lezcollitade!10

+++ src/utils/mnist/loader.h
@@ -15,14 +15,15 @@
  *  \param inputs   A pointer to the list of matrix_inputs.
  *  \param expected A pointer to the list of expected outputs.
  *  \param start    The line to start at.
- *  \param length   The number of lines to read.
+ *  \param length   The number of lines to read. If 0, data will be read until
+ *  the end.
  *  \param mode     Choose BINARY_MODE for binary files or CSV_MODE for
  *  csv files.
  *  \param inputsize    The number of inputs for 1 dataset.
  *  \param outputsize   The number of outputs for 1 dataset.
- *  \return         0 if everything worked well, -1 otherwise.
+ *  \return             The length read.
  */
-int load_mnist(const char *file, matrix_t ***inputs, matrix_t ***expected,
+size_t load_mnist(const char *file, matrix_t ***inputs, matrix_t ***expected,
     size_t start, size_t length, const char *mode, size_t inputsize,
     size_t outputsize);
